% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_SummarizedExperiment}
\alias{as_SummarizedExperiment}
\title{as_SummarizedExperiment}
\usage{
as_SummarizedExperiment(
  .data,
  assay_tables = c("Count", "ExtNPX", "ExtNPX_Corrected", "Correction", "NPX",
    "PCNormalizedNPX"),
  correction_method = c("median", "global", "none")
)
}
\arguments{
\item{.data}{\code{\link[tibble:tibble]{tibble::tibble}} of raw Olink-HT data, directly from \code{arrow::read_parquet}}

\item{assay_tables}{Which columns contain data that should be retained as a feature-by-sample matrix in the \code{assays}
slot (default: \code{Count}, \code{ExtNPX}, \code{ExtNPX_Corrected}, \code{Correction}, \code{NPX}, \code{PCNormalizedNPX})}

\item{correction_method}{Which method should be used to correct batches, \code{median}, \code{global}, or \code{none} (default: \code{median})?}
}
\value{
A SummarisedExperiment version of the Olink data
}
\description{
Convert the table from an Olink-HT parquet file into a
SummarizedExperiment object. The data must already read from disk and be in memory.
}
