% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssay-class.R
\name{assay}
\alias{assay}
\alias{assay,OlinkAssay-method}
\title{assay}
\usage{
assay(x, ...)

\S4method{assay}{OlinkAssay}(x, i = NULL, withDimnames = TRUE)
}
\arguments{
\item{x}{
    A SummarizedExperiment object or derivative.
  }

\item{...}{
    For \code{assay}, arguments in \code{...} are forwarded to \code{assays}.

    For \code{rbind}, \code{cbind}, \code{...} contains SummarizedExperiment
    objects (or derivatives) to be combined.

    For other accessors, ignored.
  }

\item{i}{name of the assay in \code{assays(x)} to retreive.}

\item{withDimnames}{retain row and column names? (Default: TRUE)}
}
\description{
Get assay data for OlinkAssay object
}
\details{
Works the same as that for \code{\link[SummarizedExperiment:assay]{SummarizedExperiment::assay}} except that
if a value for \code{i} is not provided, \code{\link{assay}} attempts to extract
the assay named (in decending order) "ExtNPX_Corrected", "ExtNPX",
"PCNormalizedNPX", "Correction", "NPX", "Count"
}
