% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{batch_correction_umap_prep}
\alias{batch_correction_umap_prep}
\title{batch_correction_umap_prep}
\usage{
batch_correction_umap_prep(
  .data,
  .col = c("ExtNPX", "ExtNPX_Corrected", "LogProtExp_Raw"),
  exclude_high_variance_assays = FALSE,
  title = NULL
)
}
\arguments{
\item{.data}{\code{\link[tibble:tibble]{tibble}} containing the level 2 data}

\item{.col}{column to use for expression data. One of \code{"ExtNPX"}, \code{"ExtNPX_Corrected"}, \code{"LogProtExp_Raw"}.}

\item{exclude_high_variance_assays}{should high variance assays be included when calculating the UMAP?
Default: FALSE}

\item{title}{title to add to plot}
}
\value{
A ggproto object (e.g. the plot object)
}
\description{
Prepare a UMAP graph from level 2 data to examine the success or failure of batch correction
}
