% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssay-class.R
\name{batchCorrection}
\alias{batchCorrection}
\title{batchCorrection}
\usage{
batchCorrection(x, ...)
}
\arguments{
\item{method}{Method to use when correcting values, either "median" or "global median".
Generally one should use the "median" methods; "global median" is more appropriate for
plates where the sample groups were not well randomized among wells.}
}
\description{
Perform batch correction using per-assay plate assay medians.
}
\details{
During import of NPX data or when creating an \code{OlinkAssay}, we calculate
the assay medians for the plate controls (those with their \code{SampleType} labeled
as "PLATE_CONTROL") and the assay medians and for all samples (\code{SampleType} ==
"SAMPLE") and store them in the \code{metadata} slot as "AssayMedians" and "GlobalMedians",
respectively. Here, we use those to calculate a correction factor:
\deqn{
 Correction_{assay} = Median_{plate} - \frac{\sum_{}^{plates}median_{plate}}{n_{plates}}
 }
\deqn{ExtNPX\_Corrected = ExtNPX - Correction_{assay}}
The corrected values are then stored as "ExtNPX_Corrected" in the assays slot
}
