% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssayGraphs.R
\name{batchCorrectionUMAP}
\alias{batchCorrectionUMAP}
\alias{batchCorrectionUMAP,OlinkAssay-method}
\title{batchCorrectionUMAP}
\usage{
batchCorrectionUMAP(x, ...)

\S4method{batchCorrectionUMAP}{OlinkAssay}(
  x,
  assay_name = c("ExtNPX", "ExtNPX_Corrected", "LogProtExp_Raw"),
  exclude_high_variance_assays = FALSE,
  title = NULL
)
}
\arguments{
\item{x}{OlinkAssay object containing batch-corrected data}

\item{assay_name}{column to use for expression data. One of \code{"ExtNPX"}, \code{"ExtNPX_Corrected"}, \code{"LogProtExp_Raw"}}

\item{exclude_high_variance_assays}{should high variance assays be included when calculating the UMAP?
Default: FALSE}

\item{title}{title to add to plot}
}
\value{
A ggproto object (e.g. the plot object)
}
\description{
Prepare a UMAP graph from level 2 data to examine the success or failure of batch correction
}
