% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssay-class.R
\name{colData}
\alias{colData}
\alias{colData,OlinkAssay-method}
\title{colData}
\usage{
colData(x, ...)

\S4method{colData}{OlinkAssay}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{
    A SummarizedExperiment object or derivative.
  }

\item{...}{
    For \code{assay}, arguments in \code{...} are forwarded to \code{assays}.

    For \code{rbind}, \code{cbind}, \code{...} contains SummarizedExperiment
    objects (or derivatives) to be combined.

    For other accessors, ignored.
  }

\item{withDimnames}{
    A \code{logical(1)}, indicating whether the dimnames of the
    SummarizedExperiment object (or derivative) should be applied (i.e.
    copied) to the extracted assays. More precisely, setting
    \code{withDimnames=FALSE} in the \emph{getter} returns the
    assays \emph{as-is} whereas setting \code{withDimnames=TRUE}
    return them with possibly modified dimnames.

    See "Top-level dimnames vs assay-level dimnames" section in
    the SummarizedExperiment vignette for more information about
    this and some examples.

    Setting \code{withDimnames=FALSE} in the \emph{setter} (\code{assays<-})
    is required when the dimnames on the supplied assays are not identical
    to the dimnames on the SummarizedExperiment object; it does not
    influence actual assignment of dimnames to assays (they're always
    stored as-is).

    Note that \preformatted{  assays(x, withDimnames=FALSE) <- assays(x, withDimnames=FALSE)}
    is guaranteed to always work and be a no-op. This is not the case
    if \code{withDimnames=TRUE} is used or if \code{withDimnames} is
    not specified.
  }
}
\description{
colData
}
