% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssay-class.R
\name{concat}
\alias{concat}
\title{concat}
\usage{
concat(x, y, ...)
}
\arguments{
\item{x}{First object to combine}

\item{y}{Second object}

\item{batch_correction_method}{method to use in performing batch correction.
See \link{medianCorrection} for details. Default: NULL}
}
\description{
Concatenate or combine two existing OlinkAssay objects.
}
\details{
Generally, an OlinkAssay object should initially be created from each
run plate separately so that the factors required for batch correction are
appropriately calculated. Using \code{concat} to combine the objects ensures that
those factors are retained and, if "median" or "global median" are passed to
\code{batch_correction_method}, used to correct the ExtNPX values.
}
