% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{ctrl_ref}
\alias{ctrl_ref}
\title{ctrl_ref}
\usage{
ctrl_ref(.data)
}
\arguments{
\item{.data}{\code{\link[tibble:tibble]{tibble::tibble}} with Olink data}
}
\description{
Calculate the default plate control batch correction
}
