% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_dfs}
\alias{.extract_dfs}
\title{.extract_dfs}
\usage{
.extract_dfs(df, filter_assay_controls = FALSE, filter_sample_controls = FALSE)
}
\description{
Remove controls for either assays, samples, or both from
the initial NPX dataframe
}
