% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_performance_report}
\alias{generate_performance_report}
\title{generate_performance_report}
\usage{
generate_performance_report(
  olink_assay,
  project_name = "Science Project",
  correction_procedure = "median",
  output_dir = here::here(),
  report_name = "assay_performance_report.pdf",
  template_file = NULL
)
}
\arguments{
\item{project_name}{title to give the report}

\item{correction_procedure}{description to use in
the report for the batch correction procedure}

\item{output_dir}{path to where the report should be output}

\item{report_name}{name to give the output report}

\item{template_file}{path to Quarto qmd file to use for report.}
}
\value{
Nothing, but writes a PDF named \code{report_name} to \code{output_dir}
}
\description{
Generate a QC performance report for a Olink HT
project.
}
\examples{
\dontrun{
generate_performance_report(
   olink_assay = here::here("olink_assay.RDS"),
   project_name = "Science Fair Project",
   correction_procedure = "Median correction",
   output_dir = "/home/user/reports",
   report_name = "assay_performance_report.pdf"
 )
}
}
