% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_umap}
\alias{graph_umap}
\title{graph_umap}
\usage{
graph_umap(
  data_umap,
  groups,
  arrow_size = 0.1,
  pt.size = 0.5,
  arrowtip_size = 2,
  cols = NULL,
  label = FALSE,
  label.size = 15
)
}
\description{
graphing pretty umap graph
}
