% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssayMethods.R
\name{level2Process}
\alias{level2Process}
\title{level2Process}
\usage{
level2Process(x, ...)
}
\description{
Calculate assay and sample QC metrics matching those as described by
the Olink HT data standard document's level 2 requirements
}
