% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{Olink_lvl1}
\alias{Olink_lvl1}
\title{olink_lvl1
Level 1 QC}
\usage{
Olink_lvl1(
  olink_files,
  proj_names,
  KitLot_info,
  PCLot_info,
  SCLot_info,
  NCLot_info
)
}
\arguments{
\item{olink_files}{Output from \code{\link[=Olink_Reader]{Olink_Reader()}}. A list with two members:
\enumerate{
\item \code{"data"}: a named list of \code{\link[tibble:tibble]{tibbles}} containing Olink data
\item \code{"manifest"}: a named list of \code{\link[tibble:tibble]{tibbles}} containing manifest data
}}

\item{proj_names}{Character vector with the projects the data is associated with}

\item{KitLot_info}{Lot number for the Olink Explore HT Kit used for the run}

\item{PCLot_info}{Lot number for Plate Control used for the Olink Explore HT run}

\item{SCLot_info}{Unknown}

\item{NCLot_info}{Lot number for Negative Control used for Olink Explore HT run}
}
\description{
olink_lvl1
Level 1 QC
}
