% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssay-class.R
\docType{class}
\name{OlinkAssay}
\alias{OlinkAssay}
\title{OlinkAssay}
\description{
An S4 class to represent data from one or more Olink HT assays
}
\section{Slots}{

\describe{
\item{\code{assays}}{a \code{\link[S4Vectors:SimpleList]{SimpleList}} of feature-by sample
\code{\link[S4Vectors:DataFrame]{DataFrames}}. Must have
"Count", "ExtNPX", "NPX" data}

\item{\code{rowData}}{\code{\link[S4Vectors:DataFrame]{DataFrame}} with feature metadata}

\item{\code{colData}}{\code{\link[S4Vectors:DataFrame]{DataFrame}} with sample metadata}

\item{\code{metadata}}{list with assay-level metadata. Data is retained per-plate/run.
Per-plate assay means are stored here
under "AssayMedians" and "GlobalMedians"}
}}

