% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssay-class.R
\name{OlinkAssayFromDisk}
\alias{OlinkAssayFromDisk}
\title{OlinkAssayFromDisk}
\usage{
OlinkAssayFromDisk(
  npx_file = NULL,
  metadata_file = NULL,
  metadata_sheet = 1,
  sample_column = "SampleID",
  project_column = "Project",
  additional_columns = NULL,
  extra_metadata = NULL,
  verbose = FALSE
)
}
\description{
OlinkAssayFromDisk
}
