% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{optimize_min_dist}
\alias{optimize_min_dist}
\title{optimize_min_dist}
\usage{
optimize_min_dist(
  .data,
  groups,
  spread,
  n_neighbor,
  min = 0.01,
  max = 0.5,
  step = 0.03
)
}
\description{
The final optimization function that is to fix the min_dist after n_neighbor and spread parameters have been fixed
}
