% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssayMethods.R
\name{retrieveAssayQC}
\alias{retrieveAssayQC}
\title{retrieveAssayQC}
\usage{
retrieveAssayQC(x, ...)
}
\arguments{
\item{x}{OlinkAssay object}
}
\description{
Extract the AssayQC results from rowData and reformat
}
