% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssayMethods.R
\name{retrieveSampleQC}
\alias{retrieveSampleQC}
\title{retrieveSampleQC}
\usage{
retrieveSampleQC(x, ...)
}
\arguments{
\item{x}{OlinkAssay object}
}
\description{
Extract the SampleQC results from colData and reformat
}
