% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssay-class.R
\name{rowData}
\alias{rowData}
\title{rowData}
\usage{
rowData(x, ...)
}
\arguments{
\item{x}{
    A SummarizedExperiment object or derivative.
  }

\item{...}{
    For \code{assay}, arguments in \code{...} are forwarded to \code{assays}.

    For \code{rbind}, \code{cbind}, \code{...} contains SummarizedExperiment
    objects (or derivatives) to be combined.

    For other accessors, ignored.
  }
}
\description{
rowData
}
