% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{setup_sdp}
\alias{setup_sdp}
\title{setup_sdp}
\usage{
setup_sdp(path)
}
\value{
Nothing, but at \code{path}, produces a structure matching
that as defined in the OlinkHT Standard Data Package, e.g.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ .
 └── SDP/
    ├── Level_1/
    │   ├── plate1.parquet
    │   ├── plate2.parquet
    │   └── manifest.xlsx
    ├── Level_2
    └── code
}\if{html}{\out{</div>}}
}
\description{
Setup a SDP hiearchical file structure
}
\examples{
\dontrun{
tmp_path <- withr::local_tempfile()
setup_sdp(path = tmp_path)
}
}
