% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{vst_to_pca}
\alias{vst_to_pca}
\title{vst_to_pca}
\usage{
vst_to_pca(data, exclude = NULL)
}
\arguments{
\item{data}{A sample-by-assay \code{\link[base:data.frame]{base::data.frame}} or \code{\link[tibble:tibble]{tibble::tibble}} with variance-stablized data}

\item{exclude}{Assays to ignore when calculating the PCA}
}
\value{
list with:
\enumerate{
\item \code{data_vst} = \code{\link[tibble:tibble]{tibble::tibble}} with assay mean, variance, predicted variance, and difference score above the predicted means
\item \code{plot} = \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} of mean vs variance with the loess function overlayed
}
}
\description{
Use variance stabilized data to detect variables/analytes with above expected variances
}
