% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{batch_correction}
\alias{batch_correction}
\title{batch_correction}
\usage{
batch_correction(data, method = c("median", "global median"))
}
\description{
Main routing function for the batch correction and normalization
}
