% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.reader_func}
\alias{.reader_func}
\title{multifile_read}
\usage{
.reader_func(
  directory,
  file_extension = c("parquet", "excel"),
  sample_column = NULL,
  project_column = NULL,
  manifest_sheet = 1
)
}
\arguments{
\item{directory}{path to where the individual Olink run data can be found}

\item{file_extension}{what kind of file should be loaded: \code{"parquet"} or \code{"excel"} (default \code{"parquet"})}
}
\value{
A named list of tibbles
}
\description{
Read multiple files into a list and name them based on the file
name minus the file extension
}
