% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.unique_rowvals}
\alias{.unique_rowvals}
\title{.unique_rowvals}
\usage{
.unique_rowvals(.data)
}
\arguments{
\item{.data}{\code{\link[tibble:tibble]{tibble::tibble}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}} with one value per column of \code{.data} with the number
distinct values in each
}
\description{
Return the unique number of values in each column
}
\examples{
data(mtcars)
.unique_rowvals(mtcars)
}
