% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_performance_report}
\alias{generate_performance_report}
\title{generate_performance_report}
\usage{
generate_performance_report(
  project_name,
  correction_procedure,
  sdp_directory,
  output_dir,
  report_name = "assay_performance_report.pdf"
)
}
\arguments{
\item{project_name}{title to give the report}

\item{correction_procedure}{description to use in
the report for the batch correction procedure}

\item{sdp_directory}{path to the directory that
includes the project \code{Level_1} and \code{Level_2}
outputs.}

\item{output_dir}{path to where the report should be output}

\item{report_name}{name to give the output report}

\item{template_dir}{path to the report markdown template}
}
\value{
Nothing, but writes a PDF named \code{report_name} to \code{output_dir}
}
\description{
Generate a QC performance report for a Olink HT
project.
}
\examples{
\dontrun{
generate_performance_report(
   project_name = "Volcano Science Project",
   correction_procedure = "median",
   sdp_directory = "/home/user/volcano_project/sdp",
   output_dir = "/home/user/reports",
   report_name = "assay_performance_report.pdf"
 )
}
}
