% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_performance_report}
\alias{generate_performance_report}
\title{generate_performance_report}
\usage{
generate_performance_report(
  project_name,
  correction_procedure,
  sdp_directory,
  report_name = "assay_performance_report.pdf",
  template_dir =
    "\\\\\\\\data/ADI/Phenotyping_Core/01_Instrumentation/Olink_HT/02_Analysis_Projects/Olink_HT_QC/Olink_HT"
)
}
\description{
generate_performance_report
}
