% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_performance_report}
\alias{generate_performance_report}
\title{generate_performance_report}
\usage{
generate_performance_report(
  project_name,
  correction_procedure,
  sdp_directory,
  output_dir,
  report_name = "assay_performance_report.pdf"
)
}
\description{
generate_performance_report
}
