% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{global_ref}
\alias{global_ref}
\title{global_ref}
\usage{
global_ref(.data)
}
\arguments{
\item{.data}{\code{\link[tibble:tibble]{tibble::tibble}} with Olink data}
}
\description{
To be used for Studies where proper randomization can be verified or where plate controls cannot be used for normalization
}
