% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_umap}
\alias{graph_umap}
\title{graph_umap}
\usage{
graph_umap(
  .data,
  groups,
  arrow_size = 0.1,
  pt.size = 0.5,
  arrowtip_size = 2,
  cols = NULL,
  label = FALSE,
  label.size = 15
)
}
\arguments{
\item{.data}{three (at least) column table
\enumerate{
\item UMAP 1
\item UMAP 2
\item Grouping variable used to color points
}}

\item{groups}{what column to suse when plotting}

\item{arrow_size}{size to give repel arrows}

\item{pt.size}{size to give arrow points}

\item{arrowtip_size}{size to draw arro tpis at}
}
\description{
graphing pretty umap graph
}
