% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_olink_data}
\alias{import_olink_data}
\title{import_olink_data}
\usage{
import_olink_data(input)
}
\arguments{
\item{input}{path to where the individual Olink run data can be found}
}
\value{
A named list of tibbles
}
\description{
Read multiple Olink output files into a list and name them based on the file name
}
\examples{
\dontrun{
import_olink_data(input = "/path/to/data")
}

}
