% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ingest_manifest}
\alias{ingest_manifest}
\title{ingest_manifest}
\usage{
ingest_manifest(directory)
}
\arguments{
\item{directory}{path to where Olink run manifest}
}
\value{
A named list of tibbles
}
\description{

}
