% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ingest_olink_data}
\alias{ingest_olink_data}
\title{ingest_olink_data}
\usage{
ingest_olink_data(directory)
}
\arguments{
\item{directory}{path to where the individual Olink run data can be found}
}
\value{
A named list of tibbles
}
\description{
Read multiple Olink output files into a list and name them based on the file name
}
