% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{median_correction}
\alias{median_correction}
\alias{median_correction.tbl_df}
\alias{median_correction.list}
\title{median_correction}
\usage{
median_correction(object, ...)

\method{median_correction}{tibble}(.data, meds)

\method{median_correction}{list}(.data, meds)

median_correction(object, ...)
}
\description{
This function takes either global or plate control median frames and applies differences

Main routing function for the batch correction and normalization
}
