% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{median_correction}
\alias{median_correction}
\title{median_correction}
\usage{
median_correction(data, meds)
}
\description{
This function takes either global or plate control median frames and applies differences
}
