% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{median_correction}
\alias{median_correction}
\alias{median_correction.tbl_df}
\alias{median_correction.list}
\title{median_correction}
\usage{
median_correction(.data, ...)

\method{median_correction}{tbl_df}(.data, medians)

\method{median_correction}{list}(.data, medians)
}
\arguments{
\item{.data}{\code{\link[tibble:tibble]{tibble}} or list of \code{tibbles} with Olink data}

\item{medians}{\code{\link[tibble:tibble]{tibble}} with three columns
\enumerate{
\item OlinkID
\item Median
\item Variance
\item PlateRef
}

Generally best to just use the output from \code{ctrl_ref()} or \code{global_ref()}.}
}
\description{
This function takes either global or plate control median frames
and applies differences
}
