% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{multifile_write}
\alias{multifile_write}
\title{multifile_write}
\usage{
multifile_write(data, file_extension = c("parquet", "csv"))
}
\arguments{
\item{data}{A named list of \link[tibble:tibble]{tibble::tibble}}

\item{file_extension}{Output type: \code{"parquet"} or \code{"csv"} (default \code{"parquet"})}
}
\value{
Nothing
}
\description{
Write multiple files into a list and name them based on the file name without file extension
}
