% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{olink_lvl2}
\alias{olink_lvl2}
\title{olink_lvl2}
\usage{
olink_lvl2(.data, multifile = TRUE, proj_dir = NULL)
}
\arguments{
\item{.data}{tibble or list of tibbles}

\item{multifile}{bool}

\item{proj_dir}{path to save to}
}
\description{
Level 2 QC Part 2
}
