% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{Olink_lvl2_prep}
\alias{Olink_lvl2_prep}
\title{Olink_lvl2_prep}
\usage{
Olink_lvl2_prep(.data)
}
\arguments{
\item{data}{a \code{\link[tibble:tibble]{tibble}} containing batch-corrected Olink data}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}}
}
\description{
Olink_lvl2_prep
}
