% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{olink_lvl2_prep}
\alias{olink_lvl2_prep}
\title{olink_lvl2_prep}
\usage{
olink_lvl2_prep(.data)
}
\arguments{
\item{.data}{a \code{\link[tibble:tibble]{tibble}} containing batch-corrected Olink data}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}}
}
\description{
olink_lvl2_prep
}
