% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{olink_qc}
\alias{olink_qc}
\title{olink_qc}
\usage{
olink_qc(.data)
}
\arguments{
\item{.data}{\code{\link[tibble:tibble]{tibble}} containing Olink data}
}
\value{
Named list with
\enumerate{
\item \code{AssayOlinkQC}: table with assay-level quality control counts
\item \code{SampleOlinkQC}: table with sample-level quality control counts
}
}
\description{
tallies the number of failed and passed assay and samples per each 96-well plate
}
