% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Olink_Reader}
\alias{Olink_Reader}
\title{Olink_Reader}
\usage{
Olink_Reader(
  directory,
  sample_column = NULL,
  project_column = NULL,
  manifest_sheet = 1
)
}
\arguments{
\item{directory}{project directory to which \code{"code"}, \code{"level 1"}, and \code{"level 2"} files should be
written and where the output from NPXExplorer and the sample manifests can be found}
}
\value{
list with \code{"data"} containing a list of \code{\link[tibble:tibble]{tibbles}}
}
\description{
Setup project directory and import Olink data and manifest
}
