% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlinkAssay-class.R
\name{OlinkAssayFromNPX}
\alias{OlinkAssayFromNPX}
\title{OlinkAssayFromNPX}
\usage{
OlinkAssayFromNPX(
  npxData,
  colData = NULL,
  assay_tables = c("Count", "ExtNPX", "PCNormalizedNPX", "NPX"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{npxData}{a \code{\link[tibble:tibble]{tibble}} with NPX data}

\item{colData}{a \code{\link[tibble:tibble]{tibble}} with extra sample metadata to store
in the \code{colData} slot.}

\item{assay_tables}{A list of columns in \code{npxData} to store
in the \code{assays} slot. (Default: "Count", "ExtNPX", "PCNormalizedNPX", "NPX")}

\item{verbose}{Should extra information about conversion progress be shown? (Default: FALSE)}
}
\value{
an \code{\link{OlinkAssay}} object
}
\description{
Convert data from an NPX parquet file to an \code{\link{OlinkAssay}} object
}
