% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{optimize_n_neighbor}
\alias{optimize_n_neighbor}
\title{optimize_n_neighbor}
\usage{
optimize_n_neighbor(
  .data,
  groups = NULL,
  spread = 10,
  min_dist = 0.1,
  min = 5,
  max = 16,
  step = 1
)
}
\description{
Optimization function for n_neighbors by setting the spread = 10, min_dist = 0.1 as default by can be changed as needed
}
