% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{optimize_spread}
\alias{optimize_spread}
\title{optimize_spread}
\usage{
optimize_spread(
  .data,
  groups,
  n_neighbor,
  min_dist = 0.1,
  min = 1,
  max = 15,
  step = 1
)
}
\description{
Optimization function for spread after n_neighbor parameter has been fixed, by default this function uses min_dist of 0.1
}
