% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{UMAP_groups}
\alias{UMAP_groups}
\title{UMAP_groups}
\usage{
UMAP_groups(
  data,
  groups,
  eps = 7,
  minPts = 10,
  arrow_size = 0.1,
  pt.size = 0.5,
  arrowtip_size = 2,
  cols = NULL,
  label = FALSE,
  label.size = 15,
  spread = 5,
  min_dist = 0.25,
  n_neighbors = NA
)
}
\description{
graph the clean reworked umap
}
